Expose concrete `*grpc.ClientConn` type rather than interface
In consul-dataplane, we intend to use a 3rd party gRPC proxy module
which accepts the concrete type rather than the interface. While we
could fork/submit a PR to the module, it doesn't make a huge amount
of difference to us which type we expose, so let's change it here
instead.