Remove custom balancer, use default pick_first (#28)
After some investigation we concluded that calling UpdateState on the ClientConn of a resolver is a sufficient cut-off point for considering the address of the connection updated. This means there is no need for a custom balancer to track sub-connections and we can simplify the watcher implementation greatly.

Since we are passing at most one address at a time, the default pick_first balancer is sufficient for our needs.